% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtInduction.R
\name{birth}
\alias{birth}
\title{birth two nodes from one bottom node (@nid) of the @tree}
\usage{
birth(
  nid,
  X,
  tree,
  fitedModel,
  predMethod,
  stumpFun,
  samplingParameters,
  rpartParas,
  nSim,
  ncores = 1
)
}
\arguments{
\item{nid}{a integer denotes the bottom node id.}

\item{X}{a dateframe includes the obsercations of covariates.}

\item{tree}{a tree (rpart object) includes the @nid bottom node.}

\item{fitedModel}{the fitted teacher model.}

\item{predMethod}{the predict function of the teacher model.}

\item{stumpFun}{a string equal to the name of stump functions. There are two of them, "stumps" and "stumpsParallel".}

\item{samplingParameters}{a list includes the sampling method name, sample size for interpretable and predictive trees and seed.}

\item{rpartParas}{a list includes the parameters of the rpart function for fitting the stump.}

\item{nSim}{a integer denotes the number of simulations.}

\item{ncores}{a integer denotes the number of CUP cores used in computation.}
}
\value{
a tree after birth.
}
\description{
birth two nodes from one bottom node (@nid) of the @tree
}
