% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingRegion.R
\name{dataRange}
\alias{dataRange}
\title{The supports of covariates}
\usage{
dataRange(X)
}
\arguments{
\item{X}{a dataframe includes all the observations of covariates.}
}
\value{
a list includes the names of covariates, the supports of continuous variables, the supports of discrete variables.
}
\description{
The supports of covariates
}
