% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.ddt}
\alias{importance.ddt}
\title{Insert the child tree's rpart.object$frame into its parent tree's rpart.object$frame.}
\usage{
importance.ddt(ddt, type)
}
\arguments{
\item{ddt}{a rpart object is the fitted tree.}

\item{type}{a strings in the vector of c("classification","regression","survival").}
}
\value{
a list includes the variable importance of the ddt.
}
\description{
Insert the child tree's rpart.object$frame into its parent tree's rpart.object$frame.
}
