% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtInduction.R
\name{inductionInterpretableTree}
\alias{inductionInterpretableTree}
\title{The induction of interpretable nodes.}
\usage{
inductionInterpretableTree(
  nids,
  type,
  X,
  fitedModel,
  predMethod,
  samplingParameters,
  nSim,
  ncores = 1
)
}
\arguments{
\item{nids}{a vector includes a set of node ids, a binary tree can be generated by automatically filling the missing nodes.}

\item{type}{a strings in the vector of c("classification","regression","survival").}

\item{X}{a dateframe includes the obsercations of covariates.}

\item{fitedModel}{the fitted teacher model.}

\item{predMethod}{the predict function of the teacher model.}

\item{samplingParameters}{a list includes the sampling method name, sample size for interpretable and predictive trees and seed.}

\item{nSim}{a integer denotes the number of simulations.}

\item{ncores}{a integer denotes the number of CUP cores used in computation.}
}
\value{
a tree includes all interpretable nodes.
}
\description{
The induction of interpretable nodes.
}
