% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtInduction.R
\name{inductionPredictionTree}
\alias{inductionPredictionTree}
\title{Induction the remaining subtrees for good prediction}
\usage{
inductionPredictionTree(
  X,
  interpretableTree,
  fitedModel,
  predMethod,
  samplingParameters,
  rpart.cp = 1e-05
)
}
\arguments{
\item{X}{a dateframe includes the obsercations of covariates.}

\item{interpretableTree}{the tree includes all interpretable nodes.}

\item{fitedModel}{the fitted teacher model.}

\item{predMethod}{the predict function of the teacher model.}

\item{samplingParameters}{a list includes the sampling method name, sample size for interpretable and predictive trees and seed.}
}
\value{
a list of subtrees and their ids, the roots are the bottom nodes of the interpretable tree.
}
\description{
Induction the remaining subtrees for good prediction
}
