% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtInduction.R
\name{kddtInductionFromInterpretable}
\alias{kddtInductionFromInterpretable}
\title{The induction of KDDT}
\usage{
kddtInductionFromInterpretable(
  X,
  y,
  interpretableTree,
  fitedModel,
  predMethod,
  samplingParameters,
  rpart.cp = 1e-05
)
}
\arguments{
\item{X}{a dateframe includes the obsercations of covariates.}

\item{y}{a vector for regression or classification, NULL for survival.}

\item{interpretableTree}{the fitted interpretable tree.}

\item{fitedModel}{the fitted teacher model.}

\item{predMethod}{the predict function of the teacher model.}

\item{samplingParameters}{a list includes the sampling method name, sample size for interpretable and predictive trees and seed.}
}
\value{
a list includes the interpretable tree and a list of subtrees.
}
\description{
The induction of KDDT
}
