% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtInduction.R
\name{kddtPredict}
\alias{kddtPredict}
\title{KDDT prediction function}
\usage{
kddtPredict(kddtModel, newX, predict_type)
}
\arguments{
\item{kddtModel}{the fitted KDDT model.}

\item{newX}{a dateframe includes the obsercations of covariates.}

\item{predict_type}{character string, c("vector", "prob", "class", "matrix"), denoting the type of predicted value returned. If the rpart object is a classification tree, then the default is to return prob predictions, a matrix whose columns are the probability of the first, second, etc. class. (This agrees with the default behavior of tree). Otherwise, a vector result is returned.}
}
\value{
a vector includes the predicted response.
}
\description{
KDDT prediction function
}
