% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStability.R
\name{plotFirstLevelStability}
\alias{plotFirstLevelStability}
\title{Plot the first-class stability of a node}
\usage{
plotFirstLevelStability(nid, firstLevelStability)
}
\arguments{
\item{nid}{a integer denotes the node id.}

\item{firstLevelStability}{a dateframe includes the pmf of covariates.}
}
\value{
a barplot show the pmf of all covariates.
}
\description{
Plot the first-class stability of a node
}
