% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStability.R
\name{plotSecondLevelStability}
\alias{plotSecondLevelStability}
\title{Plot the second-class stability of a node}
\usage{
plotSecondLevelStability(nid, secondLevelStability)
}
\arguments{
\item{nid}{a integer denotes the node id.}

\item{secondLevelStability}{a list includes all the information related to the second level stability.}
}
\value{
a density/barplot show the pdf/pmf of all covariates.
}
\description{
Plot the second-class stability of a node
}
