% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingRegion.R
\name{samplingRegion}
\alias{samplingRegion}
\title{Provide the sampling region of the node with its id.}
\usage{
samplingRegion(id, X_range, treeInfo)
}
\arguments{
\item{id}{a integer denotes the node id.}

\item{X_range}{a list includes the covariates supports.}

\item{treeInfo}{a dataframe includes the tree information.}
}
\value{
a list includes the sampling region of the node with its id.
}
\description{
Provide the sampling region of the node with its id.
}
