% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFunctions.R
\name{setRpartPara}
\alias{setRpartPara}
\title{The induction of DDT}
\usage{
setRpartPara(method, control, predict_type)
}
\arguments{
\item{method}{one of "anova", "poisson", "class" or "exp". If method is missing then the routine tries to make an intelligent guess. If y is a survival object, then method = "exp" is assumed, if y has 2 columns then method = "poisson" is assumed, if y is a factor then method = "class" is assumed, otherwise method = "anova" is assumed. It is wisest to specify the method directly, especially as more criteria may added to the function in future. Alternatively, method can be a list of functions named init, split and eval. Examples are given in the file ‘tests/usersplits.R’ in the sources, and in the vignettes ‘User Written Split Functions’.}

\item{control}{a list of options that control details of the rpart algorithm. See rpart.control.}

\item{predict_type}{character string, c("vector", "prob", "class", "matrix"), denoting the type of predicted value returned. If the rpart object is a classification tree, then the default is to return prob predictions, a matrix whose columns are the probability of the first, second, etc. class. (This agrees with the default behavior of tree). Otherwise, a vector result is returned.}
}
\value{
a list includes the parameters of rpart function.
}
\description{
The induction of DDT
}
