% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingStrategies.R
\name{setSamplingParameters}
\alias{setSamplingParameters}
\title{Set sampling parameters}
\usage{
setSamplingParameters(
  n.interpretable,
  n.predictive,
  samplingMethod = "uniform",
  intCovariates = NULL,
  exploreNodes = NULL,
  seed = 123
)
}
\arguments{
\item{n.interpretable}{the number of pseudo samples for fit the interpretable tree.}

\item{n.predictive}{the number of pseudo samples for fit the predictive subtrees.}

\item{samplingMethod}{a string denoting the sampling method, c("uniform","empirical").}

\item{intCovariates}{a vector of string includes the name of covariates whose type are integer.}

\item{exploreNodes}{a list includes the nodes need to explore more, each element of the list includes a node id and the sample size.}

\item{seed}{a integer denotes the number of random sampling}
}
\value{
a list of sampling parameters
}
\description{
Set sampling parameters
}
