% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingRegion.R
\name{setSamplingRegion}
\alias{setSamplingRegion}
\title{Set the sampling region}
\usage{
setSamplingRegion(X_range, treeInfo = NULL, id = NA)
}
\arguments{
\item{X_range}{a list includes the supports of covariates.}

\item{treeInfo}{a dataframe include the information of the DDT, default value is NULL.}

\item{id}{a integer denotes the id of the sampling region.}
}
\value{
a list includes the supports of all covariate at this node which is identified by it id.
}
\description{
Set the sampling region
}
